refactor: modernize codebase to Python 3.14
Adopt PEP 695 type parameters and `type` aliases, PEP 604 unions and PEP
758 unparenthesized except groups across the package, and raise the
toolchain floor to match (requires-python >=3.14, mypy/ruff targets,
ruff 0.8.2 -> 0.15.22).

Bugs found while validating the sweep:

* kmeans: `random_state` had silently become a no-op. Centroid init moved
  to an unseeded `np.random.default_rng()`, but `seed_all` only seeds
  NumPy's legacy global state, so five runs with the same seed produced
  five different fits and the integration test was flaky ~1 in 13. Seed a
  Generator on the instance, and give `kmeans_vectorized` a real
  `random_state` parameter to replace the lever the rewrite severed.
* inspector: `FieldInfo.type_hint` was typed `type[object]`, making
  pydantic run an `is_subclass_of` check that rejects every parameterized
  annotation (`Dict[str, Any]`, `int | None`, `Literal[...]`) -- exactly
  what this module exists to inspect. A type hint is not necessarily a
  class, so the field is now `object`.
* linear_algebra: `VectorPlotter[Vec]` shadowed the imported
  `TypeVar("Vec", bound=Vector)`, silently dropping the bound.
* hf_utils: asserted the output head is an `nn.Embedding`; it is an
  `nn.Linear` on every non-weight-tied causal LM, so it always fired.
* checkpoints: torch>=2.6 flipped `torch.load`'s `weights_only` default,
  breaking every snapshot and RNG-state resume. Snapshots store non-tensor
  state by design, so pass `weights_only=False` explicitly and document
  the trust model.
* 05_adder.ipynb still imported `get_field_annotations`, which the
  inspector rewrite removed.

Toolchain:

* Declare `pyright` in the `type` group. `make typecheck` already invoked
  it, but it was in no dependency group, so CI could never run it. Fixed
  the 27 errors it then reported -- mostly the pre-init and narrowing
  pattern already applied to `adder/main.py` but missed in its
  `main_distributed.py` and `resume.py` siblings, where the unbound
  `valid_loader`/`test_loader` were latent UnboundLocalErrors.
* Clear all 47 stale `# type: ignore` comments and split suppression
  ownership: mypy polices `# type: ignore` (`warn_unused_ignores` back
  on), pyright polices `# pyright: ignore`. Pyright's
  reportUnnecessaryTypeIgnoreComment stays off because it fires on ~34
  ignores mypy genuinely needs.
* Replace pre-commit with prek, a drop-in reimplementation that reads the
  same config (~20% faster here: 46s vs 58s). Pin the hook interpreter to
  3.14 -- both runners otherwise build hook envs on whatever they find
  (3.12/3.13), where check-ast cannot compile the PEP 758 except group.
  The hook suite now passes end to end for the first time: immutable.py
  had a stray module-level string literal that tripped check-docstring-
  first, and check-yaml skips one workflow whose 68-line commented block
  defeats prek's YAML parser (PyYAML accepts it; Actions runs it).
* Correct CONTRIBUTING.md and CLAUDE.md, which documented Python 3.12/3.9
  and several make targets and pip commands that do not exist.

make lint, security, typecheck and test all pass; mypy and pyright are
both clean; prek run --all-files is green.

+++ omnivault/machine_learning/mixture/gmm.py
@@ -320,7 +320,7 @@ class GaussianMixtureModel(BaseEstimator):
         """
         # calculate the posterior probability of each cluster for each sample
         responsibilities, marginals = self._estimate_responsibilities(X)
-        mean_marginals = np.mean(marginals)
+        mean_marginals = float(np.mean(marginals))
         return responsibilities, mean_marginals
 
     def _m_step(
@@ -492,7 +492,7 @@ class GaussianMixtureModel(BaseEstimator):
                 The reshaped Gaussian PDF values on the grid.
             """
             X_full = np.vstack((xx.ravel(), yy.ravel())).T
-            rv = multivariate_normal(mean, cov)
+            rv = multivariate_normal(mean, cov)  # pyright: ignore[reportArgumentType]  # scipy stub mistypes cov (default=1) as int; NDArray is valid
             pdf = rv.pdf(X_full)
             pdf_reshaped = pdf.reshape(xx.shape)
             return cast(NDArray[np.floating[Any]], pdf_reshaped)
